Add format option to time field
Previously there was no way to specify format for a time field, but you
can specify a format for date and datetime fields. This commit adds the
ability to pass a format option to time fields.

Because I18n uses the time locale key for for all objects with a time
[0], retain the hard-coded time field format where no format option is
passed to avoid a breaking change.

Addresses issue #1227.

[0] https://github.com/ruby-i18n/i18n/blob/0888807ab2fe4f4c8a4b780f5654a8175df61feb/lib/i18n/backend/base.rb#L82

Co-authored-by: Thom Carter <thom.carter@thoughtbot.com>